% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{effective_cor}
\alias{effective_cor}
\title{Estimate the effective correlation}
\usage{
effective_cor(design_perm, sv, target_cor, method = c("optmatch",
  "marriage"), iternum = 1000)
}
\arguments{
\item{design_perm}{A numeric design matrix whose rows are to be permuted (thus
controlling the amount by which they are correlated with the surrogate variables).
The rows index the samples and the columns index the variables.
The intercept should \emph{not} be included
(though see Section "Unestimable Components").}

\item{sv}{A matrix of surrogate variables}

\item{target_cor}{A numeric matrix of target correlations between the variables in
\code{design_perm} and the surrogate variables. The rows index the
observed covariates and the columns index the surrogate variables.
The number of columns specifies the number of surrogate variables.
Set this to \code{NULL} to indicate that \code{design_perm} and the
surrogate variables are uncorrelated.}

\item{method}{Should we use the optimal matching technique from Hansen and
Klopfer (2006) (\code{"optmatch"}) or the Gale-Shapley algorithm
for stable marriages (\code{marriage}) (Gale and Shapley, 1962)
as implemented in the matchingR package.
The \code{"optmatch"} method works almost uniformly better in practice,
but does take a lot more computational time if you have, say, 1000
samples.}

\item{iternum}{The total number of simulated correlations to consider.}
}
\description{
Will return the actual correlation between the design matrix and the
surrogate variables when you use \code{\link{permute_design}}.
}
\author{
David Gerard
}
