% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisthin.R
\name{EigenDiff}
\alias{EigenDiff}
\title{Copied code from `cate::EigenDiff` with minor changes.}
\usage{
EigenDiff(Y, rmax = min(3 * sqrt(nrow(Y)), nrow(Y), ncol(Y)),
  niter = 10)
}
\arguments{
\item{Y}{A matrix to estimate the rank.}

\item{rmax}{The maximum rank.}

\item{niter}{The maximum number of iterations.}
}
\description{
I fix an error where default `rmax` can sometimes be greater than dimension
of `Y`. I also increase the performance by only calculating the
singular values (not the singular vectors).
}
\details{
The method estimates the number of factors by the Eigenvalue Difference
Method of Onatski (2010).
}
\references{
Jingshu Wang and Qingyuan Zhao (2015). cate: High Dimensional
    Factor Analysis and Confounder Adjusted Testing and Estimation.
    R package version 1.0.4. https://CRAN.R-project.org/package=cate

    A. Onatski (2010), Determining the number of factors from empirical
    distribution of eigenvalues. The Review of Economics and Statistics 92(4).
}
\author{
David Gerard
}
