% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthin.R
\name{est_sv}
\alias{est_sv}
\title{Estimate the surrogate variables.}
\usage{
est_sv(mat, n_sv, design_fixed, design_obs, use_sva = FALSE)
}
\arguments{
\item{mat}{A matrix of counts. The rows index the genes and the columns
index the samples (as is usual in RNA-seq).}

\item{n_sv}{The number of surrogate variables.}

\item{design_fixed}{A design matrix whose rows are fixed and not permuted.
The rows index the samples and the columns index the variables.
The intercept should \emph{not} be included.}

\item{design_obs}{A matrix of observed covariates that we are NOT to be a
part of the signal generating process. Only used in estimating the
confounders if \code{use_sva = TRUE}.
The intercept should \emph{not} be included.}

\item{use_sva}{A logical. Should we use SVA using \code{design_fixed}
and \code{design_obs} to estimate the hidden covariates (\code{TRUE})
or should we just do an SVD on \code{log2(mat + 0.5)} after
subtracing row means (\code{FALSE})? Defaults to \code{FALSE}.}
}
\description{
Estimate the surrogate variables.
}
\author{
David Gerard
}
