% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthin.R
\name{fix_cor}
\alias{fix_cor}
\title{Fixes an invalid target correlation.}
\usage{
fix_cor(design_perm, target_cor, num_steps = 51)
}
\arguments{
\item{design_perm}{A design matrix whose rows are to be permuted (thus
controlling the amount by which they are correlated with the confounders).
The rows index the samples and the columns index the variables.
The intercept should \emph{not} be included.}

\item{target_cor}{A matrix of target correlation betweens the variables in
\code{design_perm} and the unobserved confounders. The rows index the
observed covariates and the columns index the unobserved confounders.
The number of columns indicates the number of hidden confounders.
Set this to \code{NULL} to indicate uncorrelated.}

\item{num_steps}{The number of steps between 0 and 1 to take in the
grid search for the shrinkage factor. The step-size would be
\code{1 / (num_steps - 1)}.}
}
\description{
Shrinks the target correlation using a uniform scaling facter so that
the overall correlation matrix is positive semi-definite.
}
\author{
David Gerard
}
