% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisthin.R
\name{poisthin}
\alias{poisthin}
\title{Apply Poisson thinning to a matrix of count data.}
\usage{
poisthin(mat, nsamp = nrow(mat), ngene = ncol(mat),
  gselect = c("max", "random", "rand_max", "custom", "mean_max"),
  gvec = NULL, skip_gene = 0, signal_fun = stats::rnorm,
  signal_params = list(mean = 0, sd = 1), prop_null = 1, alpha = 0,
  group_assign = c("half", "random", "cor"), corvec = NULL)
}
\arguments{
\item{mat}{A matrix of count data. The rows index the individuals and
the columns index the genes.}

\item{nsamp}{The number of samples to select from \code{mat}.}

\item{ngene}{The number of genes to select from \code{mat}.}

\item{gselect}{How should we select the subset of genes? Should we choose
the \code{ngene} most median expressed genes (\code{"max"}), a random sample
of the genes (\code{"random"}), a random sample of the most expressed
genes (\code{"rand_max"}), a user-provided list (\code{"custom"}), or by maximum
mean expression level (\code{"mean_max"})?
If \code{"custom"}, then \code{gvec} should be specified. Expression levels
of a gene are measured by median expression across individuals with ties broken
by mean expression.}

\item{gvec}{A logical of length \code{ncol(mat)}. A \code{TRUE} in position \eqn{i}
indicates inclusion into the smaller dataset. Hence, \code{sum(gvec)} should
equal \code{ngene}.}

\item{skip_gene}{The number of maximally expressed genes to skip.
Not used if \code{gselect = "custom"}.}

\item{signal_fun}{A function that returns the signal. This should take as
input \code{n} for the number of samples to return and then return only
a vector of samples.}

\item{signal_params}{A list of additional arguments to pass to \code{signal_fun}.}

\item{prop_null}{The proportion of genes that are null.}

\item{alpha}{If \eqn{b} is an effect and \eqn{s} is an empirical standard deviation, then
we model \eqn{b/s^\alpha} as being exchangeable.}

\item{group_assign}{How should we assign groups? Exactly half in one group
and exactly half in the other (or one less in the other if the number
of individuals is odd) (\code{"half"}), with a Bernoulli(0.5)
distribution (\code{"random"}), or correlated with latent factors
(\code{"cor"})? If \code{group_assign = "cor"}, then you have to specify
\code{corvec}.}

\item{corvec}{A vector of correlations. \code{corvec[i]} is the correlation
of the latent group assignment vector with the ith latent confounder.
Only used if \code{group_assign = "cor"}. This vector is constrained
so that \code{crossprod(corvec) < 1}. The number of latent factors
is taken to be the length of corvec.}
}
\value{
A list with the following elements:
\itemize{
 \item{\code{Y}: }{A matrix of altered counts with \code{nsamp} rows
       and \code{ngene} columns.}
 \item{\code{X}: }{A design matrix. The first column contains a vector ones (for an
       intercept term) and the second column contains an indicator for group membership.}
 \item{\code{beta}: }{The approximately true effect sizes of \eqn{log(Y) ~ X\beta}.}
 \item{\code{corassign}:}{The output from the call to \code{\link{corassign}}.
       Only returned if \code{group_assign = "cor"}.}
}
}
\description{
Given a matrix of RNA-seq counts, this function will randomly select two groups of
samples and add signal to a known proportion of the genes. This signal
is the log (base 2) effect size of the group indicator in a linear model.
The user may specify the distribution of the effects.
}
\details{
The Poisson thinning approach first randomly assigns samples to be in one of two groups. Then,
given this assignment, will Binomially sample counts with a sample size of the gene expression
counts and a probability that is a function of the effect size. For details, see
Gerard and Stephens (2017).
}
\references{
Gerard, D., & Stephens, M. (2017).
    Unifying and generalizing methods for removing unwanted variation based
    on negative controls. arXiv preprint arXiv:1705.08393.
}
\author{
David Gerard
}
