% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo.R
\name{seqgen_base}
\alias{seqgen_base}
\title{Simulate RNA-seq data from the (unrealistic) theoretical model.}
\usage{
seqgen_base(designmat, coefmat, dispvec)
}
\arguments{
\item{designmat}{A numeric design matrix. The rows index the samples
and the columns index the variables.}

\item{coefmat}{A numeric matrix of coefficients. The rows index the
genes and the columns index the variables.}

\item{dispvec}{A vector of negative binomial dispersion parameters.
Element i is the dispersion for gene i.}
}
\value{
A matrix of RNA-seq counts. The rows index the genes and the columns
    index the samples.
}
\description{
The main difference between \code{seqgen_base} and \code{seqgen_diff}
is the output format.
}
\author{
David Gerard
}
