% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theo.R
\name{seqgen_diff}
\alias{seqgen_diff}
\title{Theoretical sequence simulation for differential expression analysis.}
\usage{
seqgen_diff(designmat, coefmat, dispvec, design_sv = NULL,
  coef_sv = NULL)
}
\arguments{
\item{designmat}{A numeric design matrix. The rows index the samples
and the columns index the variables.}

\item{coefmat}{A numeric matrix of coefficients. The rows index the
genes and the columns index the variables.}

\item{dispvec}{A vector of negative binomial dispersion parameters.
Element i is the dispersion for gene i.}

\item{design_sv}{A numeric design matrix for the surrogate variables.
The rows index the samples and the columns index the surrogate
variables.}

\item{coef_sv}{A numeric matrix of coefficients for the surrogate
variables. The rows index the genes and the columns index the
variables.}
}
\description{
Generate a matrix of RNA-seq counts from a theoretical negative-binomial
model.
}
\author{
David Gerard
}
