% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthin.R
\name{thin_2group}
\alias{thin_2group}
\title{Poisson thinning in the two-group model.}
\usage{
thin_2group(mat, prop_null = 1, signal_fun = stats::rnorm,
  signal_params = list(mean = 0, sd = 1), group_prop = 0.5,
  corvec = NULL)
}
\arguments{
\item{mat}{A matrix of counts. The rows index the genes and the columns
index the samples (as is usual in RNA-seq).}

\item{prop_null}{The proportion of genes that are null.}

\item{signal_fun}{A function that returns the signal. This should take as
input \code{n} for the number of samples to return and then return only
a vector of samples. Additional parameters may be passed through
\code{signal_params}.}

\item{signal_params}{A list of additional arguments to pass to
\code{signal_fun}.}

\item{group_prop}{The proportion of individuals that are in group 1.}

\item{corvec}{A vector of target correlations. \code{corvec[i]} is the
target correlation of the latent group assignment vector with the
ith latent confounder. The default is to set this to \code{NULL},
in which case group assignment is made independently of any
unobserved confounding.}
}
\description{
Poisson thinning in the two-group model.
}
\author{
David Gerard
}
