% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthin.R
\name{thin_base}
\alias{thin_base}
\title{Basic Poisson thinning function.}
\usage{
thin_base(mat, design, coef)
}
\arguments{
\item{mat}{A matrix of counts. The rows index the genes and the columns
index the samples (as is usual in RNA-seq).}

\item{design}{A design matrix. The rows index the samples and the columns
index the variables.}

\item{coef}{A matrix of coefficients. The rows index the genes and the
columns index the samples.}
}
\description{
Given a matrix of counts, a design matrix, and a matrix
}
\author{
David Gerard
}
