% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthin.R
\name{thin_diff}
\alias{thin_diff}
\title{Poisson thinning for differential expression analysis.}
\usage{
thin_diff(mat, design_fixed = NULL, coef_fixed = NULL,
  design_perm = NULL, coef_perm = NULL, target_cor = NULL,
  use_sva = FALSE, design_obs = NULL, relative = TRUE)
}
\arguments{
\item{mat}{A matrix of counts. The rows index the genes and the columns
index the samples (as is usual in RNA-seq).}

\item{design_fixed}{A design matrix whose rows are fixed and not permuted.
The rows index the samples and the columns index the variables.
The intercept should \emph{not} be included.}

\item{coef_fixed}{The coefficients corresponding to \code{design_fixed}.
The rows index the genes and the columns index the variables.}

\item{design_perm}{A design matrix whose rows are to be permuted (thus
controlling the amount by which they are correlated with the confounders).
The rows index the samples and the columns index the variables.
The intercept should \emph{not} be included.}

\item{coef_perm}{The coefficients corresponding to \code{design_perm}.
The rows index the genes and the columns index the variables.}

\item{target_cor}{A matrix of target correlation betweens the variables in
\code{design_perm} and the unobserved confounders. The rows index the
observed covariates and the columns index the unobserved confounders.
The number of columns indicates the number of hidden confounders.
Set this to \code{NULL} to indicate uncorrelated.}

\item{use_sva}{A logical. Should we use SVA using \code{design_fixed}
and \code{design_obs} to estimate the hidden covariates (\code{TRUE})
or should we just do an SVD on \code{log2(mat + 0.5)} after
subtracing row means (\code{FALSE})? Defaults to \code{FALSE}.}

\item{design_obs}{A matrix of observed covariates that we are NOT to be a
part of the signal generating process. Only used in estimating the
confounders if \code{use_sva = TRUE}.
The intercept should \emph{not} be included.}

\item{relative}{A logical. Should we apply relative thinning (\code{TRUE})
or absolute thinning (\code{FALSE}). Only experts should change
the default.}
}
\value{
An S3 object of class \code{ThinData}. Components include some or
all of the following:
\describe{
  \item{\code{mat}}{The modified matrix of counts.}
  \item{\code{designmat}}{The design matrix, excluding an intercept term.}
  \item{\code{coefmat}}{A matrix of coefficients corresponding to \code{designmat}}
  \item{\code{sv}}{A matrix of estimated surrogate variables.}
  \item{\code{cormat}}{A matrix of target correlations between the
      surrogate variables and the permuted variables in the design matrix.}
  \item{\code{matching_var}}{A matrix of simulated variables used to
      permute the permuted components of the design matrix.}
}
}
\description{
Poisson thinning for differential expression analysis.
}
\author{
David Gerard
}
