% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthin.R
\name{thin_lib}
\alias{thin_lib}
\title{Poisson thinning for altering library size.}
\usage{
thin_lib(mat, thinlog2)
}
\arguments{
\item{mat}{A matrix of counts. The rows index the genes and the columns
index the samples (as is usual in RNA-seq).}

\item{thinlog2}{A vector. Element i is the amount to thin (on the log2 scale). For
example, a value of 0 means that we do not thin, a value of 1 means
that we thin by a factor of 2, a value of 2 means we thin by a factor
of 4, etc.}
}
\description{
Poisson thinning for altering library size.
}
\author{
David Gerard
}
