% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ThinDataToDESeqDataSet}
\alias{ThinDataToDESeqDataSet}
\title{Converts a ThinData S3 object into a DESeqDataSet S4 object.}
\usage{
ThinDataToDESeqDataSet(obj)
}
\arguments{
\item{obj}{A ThinData S3 object. This is generally output by either
\code{\link{thin_diff}}, \code{\link{thin_2group}},
\code{\link{thin_lib}}, \code{\link{thin_gene}}, or
\code{\link{thin_all}}.}
}
\value{
A \code{\link[DESeq2]{DESeqDataSet}} S4
    object. This will allow you to insert the simulated
    data directly into DESeq2.
}
\description{
The design formula in the resulting DESeqDataSet is just the sum of all
variables in \code{designmat} from the ThinData object (except the
intercept term). You should change this design formula if you want to
study other models.
}
\author{
David Gerard
}
