% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{ThinDataToSummarizedExperiment}
\alias{ThinDataToSummarizedExperiment}
\title{Converts a ThinData S3 object into a SummarizedExperiment S4 object.}
\usage{
ThinDataToSummarizedExperiment(obj)
}
\arguments{
\item{obj}{A ThinData S3 object. This is generally output by either
\code{\link{thin_diff}}, \code{\link{thin_2group}}, or
\code{\link{thin_lib}}.}
}
\value{
A \code{\link[SummarizedExperiment]{SummarizedExperiment}} S4
    object. This is often used in Bioconductor when performing
    differential expression analysis.
}
\description{
This only keeps the \code{mat} and \code{designmat} elements of
the ThinData object. Notably, the new SummarizedExperiment object
does not contain the true coefficients.
}
\author{
David Gerard
}
