% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisthin.R
\name{uncorassign}
\alias{uncorassign}
\title{Group assignment independent of anything.}
\usage{
uncorassign(n, return = c("group", "full"))
}
\arguments{
\item{n}{The sample size.}

\item{return}{Should we just return a list with just the
vector of assignment (\code{"group"})
or a list with the vector of assignments and the vector of latent
variables (\code{"full"})?}
}
\value{
A list with some or all of the following elements.
    \describe{
      \item{\code{x}}{The group assignment. \code{1L} for one group and
           \code{0L} for the other group.}
      \item{\code{w}}{The latent assignment vector (only returned if
           \code{return = "full"}). Negative corresponds to one group
           and positive corresponds to the other group.}
    }
}
\description{
Group assignment independent of anything.
}
\author{
David Gerard
}
