% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{effective_cor}
\alias{effective_cor}
\title{Estimates the effective correlation.}
\usage{
effective_cor(design_perm, sv, target_cor, calc_first = c("mean", "cor"),
  method = c("optmatch", "marriage"), iternum = 1000)
}
\arguments{
\item{design_perm}{A numeric design matrix whose rows are to be permuted (thus
controlling the amount by which they are correlated with the surrogate variables).
The rows index the samples and the columns index the variables.
The intercept should \emph{not} be included
(though see Section "Unestimable Components").}

\item{sv}{A matrix of surrogate variables}

\item{target_cor}{A numeric matrix of target correlations between the variables in
\code{design_perm} and the surrogate variables. The rows index the
observed covariates and the columns index the surrogate variables.
The number of columns specifies the number of surrogate variables.
Set this to \code{NULL} to indicate that \code{design_perm} and the
surrogate variables are uncorrelated.}

\item{calc_first}{Should we calculate the correlation of the mean
\code{design_perm} and \code{sv} (\code{calc_first = "mean"}), or should we
calculate the mean of the correlations between \code{design_perm} and
\code{sv} (\code{calc_first = "cor"})? Only \code{calc_first = "mean"}
is theoretically justified.}

\item{method}{Should we use the optimal matching technique from Hansen and
Klopfer (2006) (\code{"optmatch"}) or the Gale-Shapley algorithm
for stable marriages (\code{marriage}) (Gale and Shapley, 1962)
as implemented in the matchingR package.
The \code{"optmatch"} method works almost uniformly better in practice,
but does take a lot more computational time if you have, say, 1000
samples.}

\item{iternum}{The total number of simulated correlations to consider.}
}
\description{
Will return the estimated correlation between the design matrix and the
surrogate variables when you assign a target correlation.
}
\examples{
## Generate the design matrices and set target correlation -----------------
n <- 10
design_perm <- cbind(rep(c(0, 1), each = n / 2),
                     rep(c(0, 1), length.out = n))
sv <- matrix(rnorm(n))
target_cor <- matrix(c(0.9, 0.1), ncol = 1)

## Get estimated true correlation ------------------------------------------
## You should use a much larger iternum in practice
effective_cor(design_perm = design_perm,
              sv = sv,
              target_cor = target_cor,
              iternum = 10)

}
\author{
David Gerard
}
