% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute_design}
\alias{permute_design}
\title{Permute the design matrix so that it is approximately correlated with
the surrogate variables.}
\usage{
permute_design(design_perm, sv, target_cor, method = c("optmatch",
  "marriage"))
}
\arguments{
\item{design_perm}{A numeric design matrix whose rows are to be permuted (thus
controlling the amount by which they are correlated with the surrogate variables).
The rows index the samples and the columns index the variables.
The intercept should \emph{not} be included.}

\item{sv}{A matrix of surrogate variables}

\item{target_cor}{A numeric matrix of target correlations betweens the variables in
\code{design_perm} and the surrogate variables. The rows index the
observed covariates and the columns index the surrogate variables.
The number of columns indicates the number of surrogate variables.
Set this to \code{NULL} to indicate uncorrelated.}

\item{method}{Should we use the optimal matching technique from Hansen and
Klopfer (2006) (\code{"optmatch"}) or the Gale-Shapley algorithm
for stable marriages (\code{marriage}) (Gale and Shapley, 1962)
as implemented in the matchingR package.
The \code{"optmatch"} method works almost uniformly better in practice,
but does take a lot more computational time if you have, say, 1000
samples.}
}
\description{
Permute the design matrix so that it is approximately correlated with
the surrogate variables.
}
\references{
\itemize{
  \item{Hansen, Ben B., and Stephanie Olsen Klopfer. "Optimal full matching and related designs via network flows." Journal of computational and Graphical Statistics 15, no. 3 (2006): 609-627.}
  \item{Gale, David, and Lloyd S. Shapley. "College admissions and the stability of marriage." The American Mathematical Monthly 69, no. 1 (1962): 9-15.}
}
}
\author{
David Gerard
}
