% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisthin.R
\name{poisthin}
\alias{poisthin}
\title{Apply Poisson thinning to a matrix of count data.}
\usage{
poisthin(mat, nsamp = nrow(mat), ngene = ncol(mat), gselect = c("max",
  "random", "rand_max", "custom"), gvec = NULL, skip_gene = 0,
  signal_fun = stats::rnorm, signal_params = list(mean = 0, sd = 1),
  prop_null = 1)
}
\arguments{
\item{mat}{A matrix of count data. The rows index the individuals and
the columns index the genes.}

\item{nsamp}{The number of samples to select from \code{mat}.}

\item{ngene}{The number of genes to select from \code{mat}.}

\item{gselect}{How should we select the subset of genes? Should we choose
the \code{ngene} most expressed genes (\code{"max"}), a random sample
of the genes (\code{"random"}), a random sample of the most expressed
genes (\code{"rand_max"}), or a user-provided list (\code{"custom"})?
If \code{"custom"}, then \code{gvec} should be specified. Expression levels
of a gene are measured by median expression across individuals with ties broken
by mean expression.}

\item{gvec}{A logical of length \code{ncol(mat)}. A \code{TRUE} in position \eqn{i}
indicates inclusion into the smaller dataset. Hence, \code{sum(gvec)} should
equal \code{ngene}.}

\item{skip_gene}{The number of maximally expressed genes to skip.
Not used if \code{gselect = "custom"}.}

\item{signal_fun}{A function that returns the signal. This should take as
input \code{n} for the number of samples to return and then return only
a vector of samples.}

\item{signal_params}{A list of additional arguments to pass to \code{signal_fun}.}

\item{prop_null}{The proportion of genes that are null.}
}
\value{
A list with the following elements:
\itemize{
 \item{\code{Y}: }{A matrix of altered counts with \code{nsamp} rows
       and \code{ngene} columns.}
 \item{\code{X}: }{A design matrix. The first column contains a vector ones (for an
       intercept term) and the second column contains an indicator for group membership.}
 \item{\code{beta}: }{The approximately true effect sizes of \eqn{log(Y) ~ X\beta}.}
}
}
\description{
Apply Poisson thinning to a matrix of count data.
}
\author{
David Gerard
}

