% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genthin.R
\name{thin_2group}
\alias{thin_2group}
\title{Poisson thinning in the two-group model.}
\usage{
thin_2group(mat, prop_null = 1, signal_fun = stats::rnorm,
  signal_params = list(mean = 0, sd = 1), group_prop = 0.5,
  corvec = NULL)
}
\arguments{
\item{mat}{A numeric matrix of counts. The rows index the genes and the
columns index the samples.}

\item{prop_null}{The proportion of genes that are null.}

\item{signal_fun}{A function that returns the signal. This should take as
input \code{n} for the number of samples to return and then return only
a vector of samples. Additional parameters may be passed through
\code{signal_params}.}

\item{signal_params}{A list of additional arguments to pass to
\code{signal_fun}.}

\item{group_prop}{The proportion of individuals that are in group 1.}

\item{corvec}{A vector of target correlations. \code{corvec[i]} is the
target correlation of the latent group assignment vector with the
ith latent confounder. The default is to set this to \code{NULL},
in which case group assignment is made independently of any
unobserved confounding.}
}
\value{
A list-like S3 object of class \code{ThinData}.
Components include some or all of the following:
\describe{
  \item{\code{mat}}{The modified matrix of counts.}
  \item{\code{designmat}}{The design matrix of variables used to simulate
      signal.}
  \item{\code{coefmat}}{A matrix of coefficients corresponding to \code{designmat}}
  \item{\code{design_obs}}{Additional variables that should be included in
      your design matrix in downstream fittings. This contains at the very
      least a vector of 1's for the intercept term.}
  \item{\code{sv}}{A matrix of estimated surrogate variables. In simulation
      studies you would probably leave this out and estimate your own
      surrogate variables.}
  \item{\code{cormat}}{A matrix of target correlations between the
      surrogate variables and the permuted variables in the design matrix.}
  \item{\code{matching_var}}{A matrix of simulated variables used to
      permute \code{design_perm}.}
}
}
\description{
Given a matrix of real RNA-seq counts, this function will
randomly assign samples to one of two groups, draw
the log2-fold change in expression between two groups for each gene,
and add this signal to the RNA-seq counts matrix. The user may specify
the proportion of samples in each group, the proportion of null genes
(where the log2-fold change is 0),
and the signal function. This is a specific application of the
general Poisson thinning approach implemented in \code{\link{thin_diff}}.
}
\examples{
## Simulate data from given matrix of counts
## In practice, you would obtain Y from a real dataset, not simulate it.
set.seed(1)
nsamp <- 10
ngene <- 1000
Y <- matrix(stats::rpois(nsamp * ngene, lambda = 50), nrow = ngene)
thinout <- thin_2group(mat           = Y,
                       prop_null     = 0.9,
                       signal_fun    = stats::rexp,
                       signal_params = list(rate = 0.5))

## 90 percent of genes are null
mean(abs(thinout$coef) < 10^-6)

## Check the estimates of the log2-fold change
Ynew <- log2(t(thinout$mat + 0.5))
X    <- thinout$designmat
Bhat <- coef(lm(Ynew ~ X))["X", ]
plot(thinout$coefmat,
     Bhat,
     xlab = "log2-fold change",
     ylab = "Estimated log2-fold change")
abline(0, 1, col = 2, lwd = 2)

}
\seealso{
\describe{
  \item{\code{\link{thin_diff}}}{For the more general thinning approach.}
  \item{\code{\link{ThinDataToSummarizedExperiment}}}{For converting a
      ThinData object to a SummarizedExperiment object.}
  \item{\code{\link{ThinDataToDESeqDataSet}}}{For converting a
      ThinData object to a DESeqDataSet object.}
}
}
\author{
David Gerard
}
